+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - **All:** Add [strangelog](https://github.com/neXenio/strangelog) to maintain a changelog
+- **Selectors:** Now exposing two selectors &#x60;selectEntity(schema: string, id: string)&#x60; and &#x60;selectEntities(schema: string, ids?: string[])&#x60;
 
 ### Changed
 _No entries_